feat: contacts

+++ src/services/clients/deleteClient.service.ts
@@ -1,6 +1,7 @@
 import { AppDataSource } from "../../data-source";
 import { ClientApp } from "../../entities/clients.entities";
 import { Repository } from "typeorm";
+import { AppError } from "../../../errors";
 
 const deleteClientService = async (id: number): Promise<void> => {
   const clientRepository: Repository<ClientApp> =
@@ -11,7 +12,9 @@ const deleteClientService = async (id: number): Promise<void> => {
       id: id,
     },
   });
-
+  if (oldClientData === null) {
+    throw new AppError("cliente nao encontrado", 404);
+  }
   await clientRepository.delete(oldClientData!);
 };
 export default deleteClientService;
